requirements.txt
apache-airflow==2.10.4
pandas
requests
pyarrow
pypdf
pytest
pyspark

